+++ routes/api.php
@@ -8,7 +8,7 @@ Route::middleware('api.json')->group(function () {
 
     Route::post('/login', [AuthApiController::class, 'login']);
 
-    Route::middleware('auth:sanctum')->group(function () {
+    Route::middleware(['auth:sanctum', 'auth.session.current'])->group(function () {
         Route::get('/user', function (Request $request) {
             return $request->user();
         });